refactor: share customLogger in dev-server and preview-server
